requirements.txt
# Core Django
Django>=4.2

# Django REST Framework (API support)
djangorestframework>=3.14

# Real-time features (chat, presence)
channels>=4.0
channels-redis>=4.1

# CORS (cross-origin support for React)
django-cors-headers>=4.0

# Environment variables support
python-decouple>=3.8

# Payments (Razorpay integration)
razorpay>=1.3.0

# PostgreSQL database support (if using PostgreSQL)
psycopg2-binary>=2.9

# Email (SMTP backend with environment)
django-environ>=0.11.2

# Static/media file management (if using whitenoise)
whitenoise>=6.6.0

# Optional (Dev utilities)
ipython>=8.0

djangoframework-simplejwt<5.3.0

# ImageField Support
Pillow>=10.0

asgiref>=3.7        # required by channels
requests>=2.31      # for external API calls (e.g., Razorpay, webhooks)
uvicorn>=0.23       # ASGI server for development/production

requests